Add files via upload
Hi, I'm new to GitHub if it wasn't obvious :) 
I have a problem with my Klipper, I did a lot of tweaking and searching but can't find a solution to this problem.
If somebody could help me figure out what is wrong with Klipper that it can't open my file I'd be really happy. 
So, if I run command "Preheat", it preheats. But if I want to run the print it gives me the error that is in the .logs

Thanks in advance